pyproject.toml
[tool.poetry]
name = "google-vision-api"
version = "0.1.0"
description = "Google Vision API walkthrough"
authors = ["mattyocode <matthewoliver@live.co.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-storage = "^1.35.0"
google-cloud-vision = "^2.0.0"
numpy = "^1.19.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
